chore: Make sure bootstrap artifact is installed before returning info.
In a fresh provisioning situation, we need to do so or the info will
be empty.

We introduce a new options struct to be able to reuse the existing
installation code without printing information, since the bootstrap
artifact is a implementation detail and the user might be confused
otherwise. Error logs are still printed, so no real info should be
lost.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>