refac(http): Simplify body detection.
Previously this had not been done because I misunderstood the
`is_done()` call, and we were calling it in the wrong place. To check
for the existence of a body, we cannot use it, we need to inspect
Content-Length. But we can use it to detect that we have finished
reading the body.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>