chore(http): Fix incorrect use of `num_read` to determine payload data.
`num_read` gives you the raw bytes read from the tcp stream. For a
body with no Transfer-Encoding, this is the same as the number of
payload bytes. However, if Transfer-Encoding is enabled, for example
for chunked or compressed data, this assumption is no longer true. The
correct way to get the number of payload bytes is to look at how much
buffer space remains.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>