test: Take into account that token is not internally cached anymore.
The test was actually passing already, but this was because both
clients would initiate their requests before the round trip calling
`GetJwtToken` would finish, thereby putting the second client in the
`pending_actions_` queue of the first. This caused the number of
tokens requested to still be one, contrary to what one would expect
after disabling caching. And typically not what would happen in
production.

Fix this by forcing the two clients to be called in sequence by
stopping the event loop between them. Now `GetJwtToken` is called
twice, as expected.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>