feat(crypto): Support ED25519 signatures
ED25519 signatures are a little different. They do not work on a digest (sha),
as our other signature schemes. Instead they work on the whole message.

The signature has to be done in one run, thus it is not possible to sign a
stream. All the data has to be available from the get-go. For us this is not a
problem, since we are only signing requests, and their size is always known
beforehand.

Therefore the code needs to make an exception for the ED25519 key, as opposed to
using the generic path taken for the other keys in our signature flow.

Therefore, split the codepath in two:

1. ED25519 special case
2. A generic path, if the key is not ED25519

Ticket: None
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>