chore: Deprecate `Connectivity` settings on the client
The technical reason is that we don't maintain a cache of TCP
connections to reuse on subsequent HTTP requests as we used to have with
the golang client. We could implement it in the future if there is
demand for it.

Cancel-changelog: e6d33c6e025612e3e7120e7f5dfff6de39d31772

Ticket: MEN-6862
Changelog: Mender client does not use HTTP Keep-Alive anymore. The
related configuration option `Connectivity` is deprecated. The TCP
connection will be terminated as soon as the HTTP request/response is
finished, and every subsequent HTTP request will open a new TCP
connection.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>