fix: Close TCP socket when HTTP request/response is completed
By calling `DoCancel()` before calling the user handler in all cases
except with a successful 101 Switching Protocols response.

As the user handler will use the `response_`, remove the reset of if
(and of the `request_`) from the `DoCancel()`. This has no side effects
as these classes have noop as destructors and the objects will anyway
later be set to the new request/response in subsequent calls.

Changelog: None
Ticket: MEN-6671

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>