fix(http): Be sure to add the port number to the HOST header if nonstd
According to the spec, we need to add the HOST header with the port-number if a
nonstd port number is used.

This was discovered, as for the 1.7.x client setup, the download URL was
returned for a signed link with the port number in the signature, but we did not
add it as a part of the HOST header we set in the HTTP client.

Ticket: None
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>