fix: Make partition checking more robust in every state.
In particular, this prevents the client from erronously committing a
rootfs-image update if it's paused in the ArtifactReboot state, and
then restarted through `systemctl restart` instead of through a
reboot. Now it checks that the root partition is what we expect.

This has been a bug in the Golang client too, so we want a changelog
entry for this.

Changelog: The client no longer erronously commits a rootfs-image
artifact after being restarted using `systemctl restart` in the
`ArtifactReboot` state.

Ticket: MEN-6633

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>