chore: Add changelog for HSM change
Changelog(feat): The client no longer compiles with deprecated OpenSSL APIs.
This means that the OpenSSL HSM Engine API is deprecated for clients compiled with
OpenSSLv3, and the end-user will have to provide a v3 PKCS11 provider for the
interaction with external HSMs. For OpenSSL 1, the Engine API is still supported.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>