fix: Make all JSON parsing case insensitive
Although the JSON standard defines the keys as case sensitive, the
golang standard library implements it as case insensitive. Therefore to
avoid regressions we need to make all our JSON parsing case insensitive.

To implement it, we define a custom `nlohmann::basic_json` with a custom
`std::map` that compares the keys as lowercase. See:
* https://json.nlohmann.me/api/basic_json/basic_json/
* https://json.nlohmann.me/api/basic_json/object_t/

Ticket: MEN-6834
Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>