chore: Enable the deprecated OpenSSL Engine API
This makes some changes to enable support for the deprecated OpenSSL Engine API.

For one, we are now allowing compilation along with deprecated APIs in `common_crypto`.

For Engines (which we still support, even though they are deprecated), we load
the engine if the `SSLEngine` configuration parameter is given.

When `SSLEngine` is not set, and you are running OpenSSLv3, PKCS11 keys can be
used transparently with both Providers and files, as it is just using the URL as
the distinguisher internally. Hence we can load them both the same way.

This means quite a bit of code-change. Since the previous declaration did not
take the crypto.Args struct.

Since the tests also relied on the log output, I did unify the messages to be
`from private key` as consistently as I could.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>