build: Set MinSizeRel as the default build type
We care mostly about the size of our resulting binaries, more than we do about
performance (it is a background daemon after all).

Hence we want to build with this optimization by default.

Size of binary stripped, but not optimized:

```
-rwxr-xr-x 1 olepor users 3.0M Dec 18 16:10 ./src/mender-auth/mender-auth
-rwxr-xr-x 1 olepor users 4.1M Dec 18 16:02 ./src/mender-update/mender-update
```

Size of binary stripped and built with MinSizeRel:

```
-rwxr-xr-x 1 olepor users 1.8M Dec 18 16:10 ./src/mender-auth/mender-auth
-rwxr-xr-x 1 olepor users 2.4M Dec 18 16:09 ./src/mender-update/mender-update
```

Ticket: MEN-6911

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>