chore: Move tests out of src and remove google-test as a mandatory dependency
Put the tests in a mirrored directory structure in tests/src. This removes
the mandatory dependency on gtest, as embedded platforms such as Buildroot
and Yocto do not want nor need tests built. Also, add a new option BUILD_TESTS
which allows a user to disable building the unit tests.

Also, add a new test "no-tests" to ensure that building with the option
-DBUILD_TESTS=OFF works properly.

Signed-off-by: Adam Duskett <aduskett@gmail.com>