feat: Replace `--data` flag with `--datastore` flag.
Note in particular that we leave one instance of the old flag behind
in our tests. The flag was never tested explicitly, just implicitly
through a lot of other tests. So I just picked one at random where I
kept the old flag.

Changelog: Replace `--data` flag with `--datastore` flag. This aligns
better with the environment variables with the same names. The old
`--data` flag is still accepted for backwards compatibility, but note
that it does not have the same meaning as the `MENDER_DATA_DIR`
environment variable (hence the rename).

Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>