chore: Add first version of `TeeReader`
This object acts as a relay of some sort between one upstream source and
two or more downstream consumers. See the tests for the simple use
cases.

The next iteration will implement buffering so that a downstream reader
that joins half-way can catch up with what has already been read so far.

This is a building block for `mender-update-orchestrator` architecture,
where we will need to parse and stream the Artifact simultaneously.

Ticket: MEN-6995

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>