chore: Combine `TeeReader` with `AsyncBufferedReader`.
This is the final step for this whole chimera: make the `TeeReader` use
a buffered reader internally, so that the leafs attached half-way can
get the already streamed data.

For it we have modified the buffered readers interface to pass the
existing bytes when doing `Rewind()` and with that info we can track how
much the new leaf reader needs to catch up before we continue reading
from upstream.

It sounds more complicated that it is - look at the tests :)

Ticket: MEN-6995

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>