fix: Fix segfault when using helper method to print 0 elements
In the rare case that we would process an Artifact with 0 dependencies,
the error message printing the elements would run into a segfault.

This is a security vulnerability but setting it to no changelog as the
production code has not been released: it was developed for reuse in
other project and the code has since lived only in `master` branch.

Changelog: None
Ticket: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>