build: Remove vendored CMake projects
C++ `expected` and `optional` replacement libraries are header only and
we don't need to include their CMake projects.

The motivation of removing it is to not add a requirement on more recent
versions of CMake that are used in vendored projects (namely: `expected`
requires CMake 3.14).

This was fixed implicitly in master when all the vendored projects were
moved to a separate directory. See in the following commit that we
removed the `add_subdirectory` and moved only the `include` parts:
* https://github.com/mendersoftware/mender/commit/d9946929a63dc45e5acb6a73731ab8dd4b3b8391

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>