fix: docker-artifact-gen and mender-artifact-gen now compatible with sh
Barely anything requiring bash was used, and this makes it compatible
with systems that don't run bash. This also makes the passthrough
flags work properly, since the implementation was first created for sh.

Changelog: changed bash to sh and updated code to be compatible with sh
and the new parsing of passthrough args.

Ticket: None

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>