fix: Fix failure to call `StopBufferingAndDiscard` boundary case.
When we are exactly at the boundary between buffered reads and new
reads, in other words, when we have read all buffered data, but not
started reading the following new data, calling
`StopBufferingAndDiscard` would fail. This is because we do not detect
that we have consumed all buffered data until we have read the *next*
data after it. Combined with various uses of
`TeeReader::MakeAsyncReader`, this would lead to it not always
working.

Fix this by tracking the number of bytes we have buffered instead of
relying on encountering EOF, which enables us to set the
`rewind_consumed_` flag right away.

Independently from that, we also don't have error checking on this
condition, which would lead to the failure being silently ignored, and
therefore we would keep buffering the entire stream into memory. This
is solved in this commit, but we will also add more error checking in
the upcoming commit.

No changelog because it only affects downstream projects.

Changelog: None
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>