feat: Make it possible to compile mender-auth into mender-update.
Its commands then become available under the `auth` supercommand. Note
that global flags need to come after `auth` to be passed into it.

The feature is enabled by passing `-D EMBED_MENDER_AUTH=ON` to
cmake. This is an experimental feature, which is mainly an enabler for
porting to embedded systems. It is to be considered unsupported on
POSIX systems.

Changelog: None
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>