feat: Implement support for compiling without DBus.
This is meant as an enabler / PoC for porting to a platform without
DBus, and is not meant to be a supported configuration as is.

If the new cmake setting `MENDER_USE_DBUS` is off, then it's required
for `MENDER_EMBED_MENDER_AUTH` to be on, so that the necessary parts
are compiled into the mender-update client. The platform code for the
authentication will then use direct HTTP calls instead of calling
DBus.

Some notable highlights:

* The `Authenticator` class is made abstract and the sub classes is
  split into two platform classes, `AuthenticatorDBus` and
  `AuthenticatorHttp`. Only the bare minimum is put in the platform
  specific code, most of the logic is still shared.

* Several cmake changes. DBus libraries are no longer pulled in when
  `MENDER_USE_DBUS` is off. It also introduces several dependencies
  on mender-auth code from mender-update.

* Tests will not work in the new configuration, and I'm making no
  attempt to fix it, since this is just a PoC. I have tested the happy
  path by doing a real update, however!

Changelog: None
Ticket: MEN-7282

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>