fix: Make crypto load and use the openssl.cnf file
Libcrypto does not load the openssl.cnf file automatically.
This removes the need for manually having to export e.g.
`PKCS11_MODULE_PATH` as an env var when loading a pkcs11
interface (if it's already specified in openssl.cnf).

This also loads the config if it's changed by setting
OPENSSL_CONF.

Changelog: Title

Ticket: MEN-7353

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>