refac: Rewrite the result message code for the mender-update cli.
Add a few extra result flags to signal which operations were attempted
and failed, which helps with the messaging. This substantially
simplifies the code and makes it much more readable.

I considered switching away from the enum and use bools instead, but
the enum is still much more convenient when adding or testing for
multiple flags, so I opted to keep it.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>