fix: rootfs-image: Make it safe to roll back after `ArtifactCommit`.
Changelog(docs): Clarify in the update module protocol documentation that
going into `ArtifactRollback` is still possible after an
`ArtifactCommit`, and it must still roll back successfully. While this
may seem like it introduces new requirements into a stable protocol,
it's important to remember that this was always possible, if the
device lost power after having run the steps inside `ArtifactCommit`,
but before the device could record having done so. In this case
`ArtifactRollback` would be the next step, so this is just formalizing
this possibility and subsequent requirement.

Changelog: Title
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>