chore: add optional dynamic linking to boost::log
When building boost as a dynamic library, during linking, hundreds of
`undefined reference to `boost::log` errors occure. As the default behavior
currently is to statically link against boost::log, it is best to add a new
option that explicitly enables dynamically linking to boost::log. This new
option is called MENDER_USE_DYNAMIC_BOOST_LOG.

Signed-off-by: Adam Duskett <aduskett@gmail.com>