fix: Correclty handle error during decompression
During the decompression and therefore installation of the directory artifact the return value is not checked.

If the decompression fails (e.g. due to size constraints) the artifact install process continues in the same way it would for an successful decompression. This can leave the system in a corrupted state with a partially installed update. 

This change fixes the issue by explicitly checking the return value of the decompression process and returning an error code in case of a failure.

Changelog: None
Ticket: None

Signed-off-by: Friedrich Melchert <melchert@spectelligence.com>