feat: change generated key from RSA to ED25519
Ticket: MEN-7534
Changelog: Change the generated key from RSA to ED25519. This is
generated if there is no key provided in the configuration file, and if
and if there is no previously generated key. Existing keys won't be affected,
so this will only affect installation in new devices.
The motivation for this change is more efficient computation.

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>