fix: Don't hide out-of-space messages when streaming in update modules.
Ordinarily `Write` and `AsyncWrite` either returns the full write, or
an error. But it seems that if there is an out-of-space condition,
then a short write is returned first, then the error afterwards. The
existing code won't misbehave, but it also won't display the error,
because we need to wait for the next `Write` to get it. So do exactly
that.

This commit is a direct port of be2e3ac43edbaab from the orchestrator.

Changelog: Title
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>