fix: Limit file name length for the Artifact manifest
Limit it to 100 characters. The number is arbitrary but long enough for
the known contents of the manifest.

A manually crafted Artifact could have very long filenames and we want
to prevent the regex to run then. The `mender-artifact` tool does not
create such long filenames.

Changelog: Title

Ticket: MEN-7722

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>