chore: Require a new file for private key
The only way to be sure that no other user has access to the
file with the private key is to create a new file with proper
permissions. If there is an existing file, some other user could
have opened the file and changing permissions on it and writing
into it would give them access to the private key.

Deleting the existing file first (if any) may not be enough,
there's still a race-condition due to the window between the file
being deleted and opened for writing. Thus if creation of the
file fails, we should treat it as a failure and not write the
key out.

Ticket: MEN-7752
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek@northern.tech>