fix: Fix regression updating from Client v4 to Client v5 or newer
It is not related to the major bump of the Client, but it was reported
in Mender Hub with this suspicion so let's keep the Git commit aligned.

While working on internal improvements to support more advance use cases
in standalone mode, we introduced a new version of the standalone store.
See commit 2c1d92212.

In that commit it can be seen the code dealing with the migration (iow
an update triggered from standalone store v1 being committed with
standalone store v2): when missing the state, set the state that will
trigger the state machine into continuing into commit state.

However a later commit aiming to increase robustness of the process
added new states into the standalone state machine, and the migration
from store v1 was not taking into account, provoking a regression where
the client is not able to "continue" an update. See commit 056d9c26f.

(There are more commits in between these to, but the gist of the problem
is captured in these two).

The fix is rather simple, read it by yourself :)

Setting changelog to `None` because the bug has not been released.

Ticket: MEN-7818
Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>