fix: Resend the inventory when the device has reauthenticated
When using mutual TLS authentication the devices are authenticated by the
gateway and will be automatically authorized in the server. This
reauthentication happens without posting any error codes to the callback
when polling for new deployments. Need to instead explcitly check if the
device has reauthenticated and clear the inventory cache if that's the case.

Ticket: MEN-7820
Changelog: Title

Signed-off-by: John Olav Lund <john.olav.lund@northern.tech>