fix: retry polling with separate interval
Changelog: On failures attempting to communicate with the server when
submitting inventory or polling for deployment, a backoff mechanism
will cancel and take over for the `InventoryPollIntervalSeconds` and
`UpdatePollIntervalSeconds` intervals. The maxmimum backoff interval
is configured with `RetryPollIntervalSeconds`, and it will retry
`RetryPollCount` times. This allows mender-update to recover quicker
in case of potential race conditions, such as if mender-update starts
polling before mender-auth has generated the key on first boot.

Ticket: MEN-7790

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>