fix: Fix Mender client getting stuck after failure in sync state
From the code point, the issue was that the re-scheduling of new polls
for updates or inventory were done from the states _after_ the sync
state, so unless the state machine reached that point the new polls
would not be scheduled.

Fix by creating two new states, that just do the re-scheduling, between
idle and sync. Note that the timer(s) have now been moved to the context
object so that it can be accessed from multiple states (namely, update
polling and submit inventory states which would need to manipulate the
timer for exponential back-off retries.

Ticket: MEN-7900

Changelog: Fix issue where any error in Sync state (triggered for
example with an error in Sync_Enter state scripts) leaves the client
stuck in idle state forever and no new polls for update nor submit of
inventory would be attempted again.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>