fix: retry polling on all errors
When implementing the backoff for errors when polling for deployment or
submitting inventory, we added an exception for unauthorized errors.
After looking at how it was done in Mender Client 3, we discovered
that such an exception was not present there, and that the backoff
should be triggered for all types of errors.

Changelog: All errors on attempts to communicate with the server
are retried with an exponential backoff. This aligns the behavior
of the state machine with Mender Client 3.

Ticket: MEN-7938

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>