fix: add explicit check and handling for aborted deployments
Ticket: ME-527
Changelog: Fix an issue where a user-aborted deployment would reboot and rollback
during e.g. download. This is fixed by adding a new state event, `DeploymentAborted`,
which is posted when sending status updates to the server. Each state that sends
deployment status now explicitly checks if the deployment is aborted and handles
it accordingly.

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>