feat: Error out on comma in NO_PROXY
Instead of warning when the NO_PROXY environment variable contains a
comma, error out. The warning was easy to miss, and this is an invalid
configuration that should be reported as an error.

The error message is kept generic to apply to all mender clients that
use this shared code.

Changelog: Mender now errors if a comma is found in the NO_PROXY environment variable and explains that it only accepts space separated values.
Ticket: ME-586

Signed-off-by: Nick Anderson <nick@cmdln.org>