chore: Add an alternative implementation of KeyValueDatabase
Probably the most trivial one just using a single blob stored in
persistent storage (a file, for now, on POSIX systems) and using
basic synchronization primitives (`flock()` on the file, for
now). The goal is to be able to build the client without LMDB
and, instead, use a more portable alternative still good enough
for our trivial use-case.

Ticket: MEN-9129
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>