poc: Replace backoff algorithm with 10k retries
This is meant as an experimental change to validate a use case of the
client in poor networks, where a typical download will need more than 10
retries.

This commit replaces the backoff mechanism for the download retries with
a fixed interval of 30 seconds to retry and a maximum of 10k retries.

Ticket: ME-589

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>